fix(tests): resolve compilation errors and update test analysis
Fix compilation errors in integration tests caused by IFD parser API
change from 2-tuple to 3-tuple return type (adding field_type). Update
all tuple destructuring patterns across test files. Remove unnecessary
type casts in operations.rs. Expand codemachine prompts with detailed
test failure analysis for PNG, PDF, MP4, TIFF, and XMP parsers.

Changes:
- Update tuple destructuring in jpeg_tests.rs, tiff_tests.rs
- Update write test patterns in jpeg_write_tests.rs, tiff_write_tests.rs
- Fix clippy warning for unnecessary i32 casts in operations.rs
- Expand code_fallback.md with comprehensive test failure breakdown
- Update context.md with integration test plan details

+++ src/core/operations.rs
@@ -508,7 +508,7 @@ fn raw_bytes_to_tag_value(bytes: &[u8], field_type: u16, byte_order: ByteOrder) 
                     ByteOrder::LittleEndian => i32::from_le_bytes([bytes[4], bytes[5], bytes[6], bytes[7]]),
                     ByteOrder::BigEndian => i32::from_be_bytes([bytes[4], bytes[5], bytes[6], bytes[7]]),
                 };
-                return TagValue::new_rational(numerator as i32, denominator as i32);
+                return TagValue::new_rational(numerator, denominator);
             }
 
             // ASCII (type 2): null-terminated string
